// src/gql/graphql.ts
export type ArticleItemFragment = { slug: string, title: string, text: string, createdat: unknown, link?: string | null };